tui: extract app-server adapter module
